refactor: move the sql_helper to core to avoid layer leakage
